app/requirements.txt
streamlit
numpy
tensorflow
pillow
streamlit-webrtc
av
mlflow